requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
pydantic==1.8.1
python-dotenv==0.17.0
requests==2.25.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
